DCAC-78: Rationalise integration tests.
